rename Registry
