+++ packages/common/src/index.ts
@@ -1,7 +1,7 @@
 // import './test';
 
 export * from './utils';
-export * from './Registry2';
+export * from './Registry';
 
 export * from './ConfiguFile';
 export * from './CfguFile';
